requirements.txt
streamlit
google-genai
feedparser
